Mostly work on power, regulators and distribution.  Also more BOM cleanup.
All of the bypass capacitors are now spec'd and in the DigiKey BOM. TVS zeners are now
spec'd.  Folded together some resistor values and got rid of some 1206 parts that no longer
need that much power rating at the lower supplies.

A affecting all of the input channel sheets is that we now distribute filtered V+, V- and VREF
via sheet ports.  The V+, V- filter is on input_channel.sch and the VREF filter is still in
adc.sch, but is distributed out to the other sheets that need it.

bist_buffer.sch: added buffer for built-in self-test

board_connector.sch: put the input filter for the analog supplies here.

power.sch: added regulators for the supplies.

reference_buffer.sch:
changed components to match those chosen for the V+, V- regulators and added
current steering diodes to bootstrap the reference buffer supply.  This is
necessary because the analog supplies are now derived from VREF.

*.wxsch:
Updates to spice models related to work on supplies.