Cleanup to get ready to start board layout. I had to get rid off all the bus signals, since in Kicad busses don't work that way.  Added a couple testpoints; many more needed.
Work on component placement in PCB.  I changed DGND to plain GND and "protective earth" to
"earth" (for chassis ground).  Chassis ground is now connected on the board only via a capacitor.
I had to fix the connector footprint to match the pin numbers on the symbol (there was a case
difference).