Minor tweaks to schematic, and ongoing progress on PCB layout.
Changed BIST buffer to run off of CHAN1_V+, not the global V+.  This caused the CHAN_V+ net name to change.
On power.sch, added bypass caps for the VREF inputs to the V+/V- regulators.  Also, the BIST input lines now
have an RC filter, not just a resistor.