Various tweaks to schematic which help in the PCB view, mainly getting informative net names. Also unscrambled the + and - paths in the input channel.  This avoided a cross-over before the first stage, and an unintended inversion, but mostly just made net naming less confusing.
Getting better net names made it easier to review the PCB to see if it made sense.  I checked
the ICs on the PCB against the datasheet pinout, and while doing this fixed an actual bug where
I had BIST-+ and BIST+-swapped.

I made some tweaks on the PCB to try to equalize the copper area on the input nets so as to
match the capacitance.  I tweaked the top-side ground pours somewhat, adding in the regulator
area, removing around the high speed traces.

Some changes from gerber viewing.  I reduced the solder mask clearances so that we get mask
web between the SOIC and DFN pads.  I reduced the pad size of the big via from 1.2 to 1.0 mm.